ml_service/requirements.txt
Flask==2.2.2
Werkzeug==2.3.8
torch
torchvision
numpy
Pillow